ai-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
numpy==2.2.4
Pillow==11.1.0
opencv-python==4.11.0.86
